pyproject.toml
[tool.poetry]
name = "doggobot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.2"
colorthief = "^0.2.1"
timeago = "^1.0.15"
jishaku = "^1.20.0"
tasks = "^2.8.0"
flask = "^2.0.1"
urbandict = "^0.6.1"
ping3 = "^2.7.0"
discord-buttons-plugin = "^0.1.2"
dhooks = "^1.1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
